feat: Prevent proof submission when not required
